update mos draw
