+++ pdkprims/src/contact.rs
@@ -9,7 +9,7 @@ use layout21::utils::Ptr;
 use serde::{Deserialize, Serialize};
 
 use crate::config::Int;
-use crate::geometry::CoarseDirection;
+use crate::geometry::{expand_box, CoarseDirection};
 use crate::{config::Uint, Pdk};
 
 #[derive(Debug, Clone, Eq, PartialEq, Hash, Serialize, Deserialize, derive_builder::Builder)]
@@ -230,22 +230,5 @@ impl Pdk {
             cols: params.cols,
             bboxes: bbox_map,
         }
-    }
-}
-
-fn expand_box(b: &mut Rect, dist: Int) {
-    assert!(b.p0.x <= b.p1.x);
-    assert!(b.p0.y <= b.p1.y);
-
-    b.p0.x -= dist;
-    b.p1.x += dist;
-    b.p0.y -= dist;
-    b.p1.y += dist;
-}
-
-fn rect_from_bbox(bbox: &BoundBox) -> Rect {
-    Rect {
-        p0: bbox.p0.clone(),
-        p1: bbox.p1.clone(),
     }
 }
