+++ pdkprims/src/geometry.rs
@@ -1,5 +1,6 @@
 use std::fmt::Display;
 
+use layout21::raw::{BoundBox, Rect};
 use serde::{Deserialize, Serialize};
 
 /// A direction: horizontal or vertical.
@@ -30,5 +31,22 @@ impl CoarseDirection {
             Self::Horizontal => "h",
             Self::Vertical => "v",
         }
+    }
+}
+
+pub fn expand_box(b: &mut Rect, dist: crate::config::Int) {
+    assert!(b.p0.x <= b.p1.x);
+    assert!(b.p0.y <= b.p1.y);
+
+    b.p0.x -= dist;
+    b.p1.x += dist;
+    b.p0.y -= dist;
+    b.p1.y += dist;
+}
+
+fn rect_from_bbox(bbox: &BoundBox) -> Rect {
+    Rect {
+        p0: bbox.p0.clone(),
+        p1: bbox.p1.clone(),
     }
 }
